Implement Send + Sync for LedCanvas (#47)
Since led_matrix_swap_on_vsync() will block until the next vsync,
implementing the Send + Sync traits for LedCanvas will allow the canvas
to be sent between a render thread and a draw thread. Since the
underlying canvas is created with CreateFrameCanvas(), which allocates
the canvas memory on the heap, it is safe to share between threads.